Extract cue_flac test module into cue_flac/tests.rs
Move the inline `#[cfg(test)] mod tests { … }` block out of
bae-core/src/cue_flac.rs into a sibling child-module file,
bae-core/src/cue_flac/tests.rs, leaving only `#[cfg(test)] mod tests;` behind.
Pure relocation: no test code changed (the body is token-identical, just
reindented to module top level by rustfmt). The child module still sees the
parent's private items through `use super::*;`.
